requirements.txt
altgraph==0.17.2
future==0.18.2
importlib-metadata==4.8.1
numpy==1.21.2
opencv-python==4.5.3.56
pefile==2021.9.3
Pillow==9.2.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pywin32-ctypes==0.2.0
typing-extensions==3.10.0.2
zipp==3.5.0
